Add Error type and organise modules
Moving the waiting room into its own module to separate the boundary
that it's a separate entity to request.

This makes it easier to define the Error types locally and replace the
() errors with these domain specific errors.
